wip: updated firebase

+++ routes/travelRoutes.js
@@ -34,6 +34,14 @@ router.post("/locations", isAuthenticated, async (req, res, next) => {
 // update route
 router.put('/locations/:id', isAuthenticated, async (req, res, next) => {
   try {
+      const location = await Travel.findById(req.params.id);
+      if (!location) return res.status(404).send('Location not found');
+
+      // Check if the UID in the entry matches the UID from the token
+      if (location.uid !== req.user.uid) {
+          return res.status(403).send('Permission denied');
+      }
+
       const updatedData = await Travel.findByIdAndUpdate(req.params.id, req.body, { new: true });
       res.json(updatedData);
   } catch (error) {
@@ -42,13 +50,23 @@ router.put('/locations/:id', isAuthenticated, async (req, res, next) => {
 });
 
 
+
 // delete
 router.delete("/locations/:id", isAuthenticated, async (req, res, next) => {
   try {
+      const location = await Travel.findById(req.params.id);
+      if (!location) return res.status(404).send('Location not found');
+
+      // Check if the UID in the entry matches the UID from the token
+      if (location.uid !== req.user.uid) {
+          return res.status(403).send('Permission denied');
+      }
+
       res.json(await Travel.findByIdAndDelete(req.params.id));
   } catch (error) {
       res.status(400).json(error);
   }
 });
+
 
 module.exports = router;
